app/pyproject.toml
[tool.poetry]
name = "noth-api"
version = "0.1.0"
description = ""
authors = ["Romain Grente <romain.grente@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
SQLAlchemy = ">=1.4,<1.4.23"
itsdangerous = "^2.0.1"
Authlib = "^0.15.4"
vonage = "^2.5.5"
alembic = "^1.6.5"
PyJWT = "^2.1.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
